# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARKLAND Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface SessionUser {
  /** @nullable */
  steamId?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin: boolean;
  points: number;
  authenticated: boolean;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
  productCount: number;
}

export interface Product {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  price: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categorySlug?: string | null;
  /** @nullable */
  blueprint?: string | null;
  quantity: number;
  quality: number;
  forceBlueprint: boolean;
  isFeatured: boolean;
  isActive: boolean;
  createdAt: string;
}

export interface ProductPage {
  items: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ProductInput {
  slug: string;
  name: string;
  description?: string;
  imageUrl?: string;
  price: number;
  categoryId?: number;
  blueprint?: string;
  quantity: number;
  quality?: number;
  forceBlueprint?: boolean;
  isFeatured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  imageUrl?: string;
  price?: number;
  /** @nullable */
  categoryId?: number | null;
  blueprint?: string;
  quantity?: number;
  quality?: number;
  forceBlueprint?: boolean;
  isFeatured?: boolean;
  isActive?: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  delivered: 'delivered',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  steamId: string;
  /** @nullable */
  displayName?: string | null;
  productId: number;
  productName: string;
  quantity: number;
  pointsPaid: number;
  status: OrderStatus;
  /** @nullable */
  paymentId?: string | null;
  /** @nullable */
  paymentUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderPage {
  items: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface OrderInput {
  productId: number;
  quantity?: number;
}

export interface OrderCreated {
  order: Order;
  /** @nullable */
  paymentUrl?: string | null;
  usePoints: boolean;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  delivered: 'delivered',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type WebhookPayloadData = {
  id?: string;
};

export interface WebhookPayload {
  type?: string;
  data?: WebhookPayloadData;
}

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  totalOrders: number;
  totalDelivered: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
};

export type AdminListProductsParams = {
category?: string;
page?: number;
};

export type AdminListOrdersParams = {
status?: string;
page?: number;
};