# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ARKLAND Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: products
    description: Store products
  - name: categories
    description: Product categories
  - name: orders
    description: Orders
  - name: payments
    description: Payment webhooks
  - name: stats
    description: Store statistics
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user (null if not logged in)
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: List featured/promoted products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List current user's orders (auth required)
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order and receive payment URL (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order with payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get single order by id (must belong to current user or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/webhook:
    post:
      operationId: paymentsWebhook
      tags: [payments]
      summary: MercadoPago IPN/webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /store/stats:
    get:
      operationId: getStoreStats
      tags: [stats]
      summary: Public store stats
      responses:
        "200":
          description: Store stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin — list all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Products page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Admin — create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin — update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Admin — delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin — list all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Orders page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"

  /admin/orders/{id}:
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      summary: Admin — update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SessionUser:
      type: object
      properties:
        steamId:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        points:
          type: integer
        authenticated:
          type: boolean
      required:
        - authenticated
        - isAdmin
        - points

    Category:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
        productCount:
          type: integer
      required:
        - id
        - slug
        - name
        - sortOrder
        - productCount

    Product:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        price:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categorySlug:
          type: ["string", "null"]
        blueprint:
          type: ["string", "null"]
        quantity:
          type: integer
        quality:
          type: number
        forceBlueprint:
          type: boolean
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - name
        - price
        - quantity
        - quality
        - forceBlueprint
        - isFeatured
        - isActive
        - createdAt

    ProductPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - totalPages

    ProductInput:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: integer
        categoryId:
          type: integer
        blueprint:
          type: string
        quantity:
          type: integer
        quality:
          type: number
        forceBlueprint:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - slug
        - name
        - price
        - quantity

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: integer
        categoryId:
          type: ["integer", "null"]
        blueprint:
          type: string
        quantity:
          type: integer
        quality:
          type: number
        forceBlueprint:
          type: boolean
        isFeatured:
          type: boolean
        isActive:
          type: boolean

    Order:
      type: object
      properties:
        id:
          type: integer
        steamId:
          type: string
        displayName:
          type: ["string", "null"]
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        pointsPaid:
          type: integer
        status:
          type: string
          enum: [pending, paid, delivered, failed, refunded]
        paymentId:
          type: ["string", "null"]
        paymentUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - steamId
        - productId
        - productName
        - quantity
        - pointsPaid
        - status
        - createdAt
        - updatedAt

    OrderPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - totalPages

    OrderInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          default: 1
      required:
        - productId

    OrderCreated:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        paymentUrl:
          type: ["string", "null"]
        usePoints:
          type: boolean
      required:
        - order
        - usePoints

    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, delivered, failed, refunded]
      required:
        - status

    WebhookPayload:
      type: object
      properties:
        type:
          type: string
        data:
          type: object
          properties:
            id:
              type: string

    StoreStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        totalOrders:
          type: integer
        totalDelivered:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - totalOrders
        - totalDelivered