# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARKLAND Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  steamId: string;
  /** @nullable */
  displayName?: string | null;
  productId: number;
  productName: string;
  quantity: number;
  pointsPaid: number;
  status: OrderStatus;
  /** @nullable */
  paymentId?: string | null;
  /** @nullable */
  paymentUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}