# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARKLAND Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  price: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categorySlug?: string | null;
  /** @nullable */
  blueprint?: string | null;
  quantity: number;
  quality: number;
  forceBlueprint: boolean;
  isFeatured: boolean;
  isActive: boolean;
  createdAt: string;
}