# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARKLAND Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  slug: string;
  name: string;
  description?: string;
  imageUrl?: string;
  price: number;
  categoryId?: number;
  blueprint?: string;
  quantity: number;
  quality?: number;
  forceBlueprint?: boolean;
  isFeatured?: boolean;
}