# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARKLAND Store API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  /** @nullable */
  steamId?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin: boolean;
  points: number;
  authenticated: boolean;
}